% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerUtility-package.R
\docType{package}
\name{cerUtility-package}
\alias{cerUtility}
\alias{cerUtility-package}
\title{cerUtility: Utilities for CER Packages}
\description{
Utilities for CER Packages.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
