% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EbmUtility-package.R
\docType{package}
\name{EbmUtility-package}
\alias{EbmUtility}
\alias{EbmUtility-package}
\title{EbmUtility: Utilities for EBM Packages}
\description{
Utilities for EBM Packages.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
